Create maven.yml
Added the Java Create Maven from the action tab to the project. It should run GitHub Action is called in two cases: when a push is made to the main branch and when a pull-request is created to the main branch. when a pull-request is created in the main branch.